requirements.txt
avro==1.12.0
beautifulsoup4==4.12.3
pyarrow==17.0.0
PyYAML==6.0.2
Requests==2.32.3
tqdm==4.66.5
